chore(api): Move chronicle related models to a chronicle folder
