+++ api/src/models/api/stardust/chronicle/IAddressBalanceResponse.ts
@@ -1,4 +1,4 @@
-import { IResponse } from "../IResponse";
+import { IResponse } from "../../IResponse";
 
 export interface IAddressBalanceResponse extends IResponse {
     /**
