+++ api/src/models/api/stardust/chronicle/ITransactionHistoryDownloadResponse.ts
