+++ api/src/routes/stardust/transactionhistory/download/post.ts
@@ -2,8 +2,8 @@ import JSZip from "jszip";
 import { ServiceFactory } from "../../../../factories/serviceFactory";
 import logger from "../../../../logger";
 import { IDataResponse } from "../../../../models/api/IDataResponse";
-import { ITransactionHistoryDownloadBody } from "../../../../models/api/stardust/ITransactionHistoryDownloadBody";
-import { ITransactionHistoryRequest } from "../../../../models/api/stardust/ITransactionHistoryRequest";
+import { ITransactionHistoryDownloadBody } from "../../../../models/api/stardust/chronicle/ITransactionHistoryDownloadBody";
+import { ITransactionHistoryRequest } from "../../../../models/api/stardust/chronicle/ITransactionHistoryRequest";
 import { IConfiguration } from "../../../../models/configuration/IConfiguration";
 import { STARDUST } from "../../../../models/db/protocolVersion";
 import { NetworkService } from "../../../../services/networkService";
